backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
scikit-learn==1.4.0
numpy==1.26.4
pydantic[email]==2.6.1
python-multipart==0.0.9
